Add new ministerial interactions dashboard pipeline
- Move all dashboard pipelines to a single file
