+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-05-06
+
+- Move dashboard DAGs to a single file `dashboard_pipelines.py`
+- Add new dashboard DAG for ministerial interactions 
+
 ## 2020-05-05
 
 ### Changed
